Refactor, alfabetize, remove doublets
