+++ members/admin/activityinvite_admin.py
@@ -1,0 +1,103 @@
+from django import forms
+from django.contrib import admin
+from django.db.models.functions import Lower
+from django.utils import timezone
+
+from members.models import (
+    Activity,
+    ActivityInvite,
+    AdminUserInformation,
+    Department,
+    Person,
+)
+
+
+class ActivityInviteAdminForm(forms.ModelForm):
+    class Meta:
+        model = ActivityInvite
+        exclude = []
+
+    def __init__(self, *args, **kwds):
+        super(ActivityInviteAdminForm, self).__init__(*args, **kwds)
+        self.fields["person"].queryset = Person.objects.order_by(Lower("name"))
+
+
+class ActivivtyInviteActivityListFilter(admin.SimpleListFilter):
+    title = "Aktiviteter"
+    parameter_name = "activity"
+
+    def lookups(self, request, model_admin):
+        activitys = []
+        for activity in Activity.objects.filter(
+            department__in=AdminUserInformation.get_departments_admin(request.user)
+        ).order_by("department__name"):
+            activitys.append((str(activity.pk), activity))
+
+        return activitys
+
+    def queryset(self, request, queryset):
+        if self.value() is None:
+            return queryset
+        else:
+            return queryset.filter(activity__pk=self.value())
+
+
+class ActivityInviteAdmin(admin.ModelAdmin):
+    list_display = (
+        "person",
+        "activity",
+        "invite_dtm",
+        "person_age_years",
+        "person_zipcode",
+        "rejected_dtm",
+    )
+    list_filter = (ActivivtyInviteActivityListFilter,)
+    search_fields = ("person__name",)
+    list_display_links = None
+    form = ActivityInviteAdminForm
+
+    # Only show invitation to own activities
+    def get_queryset(self, request):
+        qs = super(ActivityInviteAdmin, self).get_queryset(request)
+        if request.user.is_superuser:
+            return qs
+        return qs.filter(activity__department__adminuserinformation__user=request.user)
+
+    fieldsets = (
+        (
+            None,
+            {
+                "description": '<p>Invitationer til en aktivitet laves nemmere via "person" oversigten. Gå derind og filtrer efter f.eks. børn på venteliste til din afdeling og sorter efter opskrivningsdato, eller filter medlemmer på forrige sæson. Herefter kan du vælge de personer på listen, du ønsker at invitere og vælge "Inviter alle valgte til en aktivitet" fra rullemenuen foroven.</p>',
+                "fields": (
+                    "person",
+                    "activity",
+                    "invite_dtm",
+                    "expire_dtm",
+                    "rejected_dtm",
+                ),
+            },
+        ),
+    )
+
+    # Limit the activity possible to invite to: Not finished and belonging to user
+    def formfield_for_foreignkey(self, db_field, request, **kwargs):
+        if db_field.name == "activity" and not request.user.is_superuser:
+            departments = Department.objects.filter(
+                adminuserinformation__user=request.user
+            )
+            kwargs["queryset"] = Activity.objects.filter(
+                end_date__gt=timezone.now(), department__in=departments
+            )
+        return super(ActivityInviteAdmin, self).formfield_for_foreignkey(
+            db_field, request, **kwargs
+        )
+
+    def person_age_years(self, item):
+        return item.person.age_years()
+
+    person_age_years.short_description = "Alder"
+
+    def person_zipcode(self, item):
+        return item.person.zipcode
+
+    person_zipcode.short_description = "Postnummer"
