+++ members/admin/person_admin.py
@@ -2,14 +2,14 @@ from datetime import timedelta
 
 from django import forms
 from django.contrib import admin
+from django.contrib import messages
+from django.contrib.admin.widgets import AdminDateWidget
 from django.db import transaction
 from django.db.models import Q
 from django.http import HttpResponse
 from django.utils import timezone
 from django.utils.html import format_html
 from django.shortcuts import render
-from django.contrib.admin.widgets import AdminDateWidget
-from django.contrib import messages
 
 from members.models import (
     Activity,
@@ -19,17 +19,17 @@ from members.models import (
 )
 
 from .person_admin_filters import (
-    VolunteerListFilter,
-    PersonWaitinglistListFilter,
     PersonInvitedListFilter,
     PersonParticipantListFilter,
+    PersonWaitinglistListFilter,
+    VolunteerListFilter,
 )
 
 from .inlines import (
+    ActivityInviteInline,
     MemberInline,
     PaymentInline,
     VolunteerInline,
-    ActivityInviteInline,
     WaitingListInline,
 )
