+++ members/admin/person_admin_filters.py
@@ -1,4 +1,5 @@
 from django.contrib import admin
+
 from members.models import Activity, AdminUserInformation
